Move event closure/reopening controllers from applications.views to applications.helpers, add first_speaker and submitted_applications fields to EventDetailSerializer
