+++ src/applications/views.py
@@ -10,20 +10,22 @@ from rest_framework.mixins import (
 )
 from rest_framework.viewsets import GenericViewSet
 
+from .helpers import (
+    EventClosureController,
+    EventReopeningController,
+    create_notification_settings,
+)
 from .models import Application, NotificationSettings
 from .serializers import (
     ApplicationCreateAnonymousSerializer,
     ApplicationCreateAuthorizedSerializer,
     NotificationSettingsSerializer,
 )
-from .utils_db_write import create_notification_settings
 from api.loggers import logger
 from api.permissions import IsAuthorOrCreateOnly
-from events.models import Event
 from users.models import Specialization
 
 
-# TODO: Если заявка отменена авторизованным юзером, то открывать регистрацию снова.
 # TODO: Возвращать статус 200 и response body в случае удаления объекта
 class ApplicationViewSet(CreateModelMixin, DestroyModelMixin, GenericViewSet):
     """ViewSet to create and delete applications for participation in events."""
@@ -72,128 +74,11 @@ class ApplicationViewSet(CreateModelMixin, DestroyModelMixin, GenericViewSet):
             )
         user.save()
 
-    @staticmethod
-    def check_offline_event_limit(event: Event) -> bool:
-        """Checks if the participant limit for the offline event is reached."""
-        return (
-            event.format == Event.FORMAT_OFFLINE
-            and event.participant_offline_limit
-            and event.applications.filter(format=Event.FORMAT_OFFLINE).count()
-            >= event.participant_offline_limit
-        )
-
-    @staticmethod
-    def check_online_event_limit(event: Event) -> bool:
-        """Checks if the participant limit for the online event is reached."""
-        return (
-            event.format == Event.FORMAT_ONLINE
-            and event.participant_online_limit
-            and event.applications.filter(format=Event.FORMAT_ONLINE).count()
-            >= event.participant_online_limit
-        )
-
-    @staticmethod
-    def check_hybrid_event_offline_reached_online_not_reached(event: Event) -> bool:
-        """
-        Checks that the offline participant limit has been reached and
-        the online participant limit has not been reached.
-        """
-        offline_limit_reached: bool = (
-            event.format == Event.FORMAT_HYBRID
-            and event.participant_offline_limit
-            and event.applications.filter(format=Event.FORMAT_OFFLINE).count()
-            >= event.participant_offline_limit
-        )
-        online_limit_not_reached: bool = (
-            not event.participant_online_limit
-            or event.applications.filter(format=Event.FORMAT_ONLINE).count()
-            < event.participant_online_limit
-        )
-        return offline_limit_reached and online_limit_not_reached
-
-    @staticmethod
-    def check_hybrid_event_online_reached_offline_not_reached(event: Event) -> bool:
-        """
-        Checks that the online participant limit has been reached and
-        the offline participant limit has not been reached.
-        """
-        online_limit_reached: bool = (
-            event.format == Event.FORMAT_HYBRID
-            and event.participant_online_limit
-            and event.applications.filter(format=Event.FORMAT_ONLINE).count()
-            >= event.participant_online_limit
-        )
-        offline_limit_not_reached: bool = (
-            not event.participant_offline_limit
-            or event.applications.filter(format=Event.FORMAT_OFFLINE).count()
-            < event.participant_offline_limit
-        )
-        return online_limit_reached and offline_limit_not_reached
-
-    @staticmethod
-    def check_hybrid_event_offline_limit(event: Event) -> bool:
-        """Checks if the participant offline limit for the hybrid event is reached."""
-        return (
-            event.format == Event.FORMAT_HYBRID
-            and event.participant_offline_limit
-            and event.applications.filter(format=Event.FORMAT_OFFLINE).count()
-            >= event.participant_offline_limit
-        )
-
-    @staticmethod
-    def check_hybrid_event_online_limit(event: Event) -> bool:
-        """Checks if the participant online limit for the hybrid event is reached."""
-        return (
-            event.format == Event.FORMAT_HYBRID
-            and event.participant_online_limit
-            and event.applications.filter(format=Event.FORMAT_ONLINE).count()
-            >= event.participant_online_limit
-        )
-
-    @staticmethod
-    def check_event_limits_and_close_registration(event: Event) -> None:
-        """
-        Checks event participant limits and closes registration if limits are reached.
-        """
-        if ApplicationViewSet.check_hybrid_event_offline_reached_online_not_reached(
-            event
-        ):
-            event.status = Event.STATUS_OFFLINE_CLOSED
-            logger.debug(
-                f"The status of event {event} was changed to "
-                f"{Event.STATUS_OFFLINE_CLOSED}"
-            )
-        elif ApplicationViewSet.check_hybrid_event_online_reached_offline_not_reached(
-            event
-        ):
-            event.status = Event.STATUS_ONLINE_CLOSED
-            logger.debug(
-                f"The status of event {event} was changed to "
-                f"{Event.STATUS_ONLINE_CLOSED}"
-            )
-        elif (
-            ApplicationViewSet.check_offline_event_limit(event)
-            or ApplicationViewSet.check_online_event_limit(event)
-            or (
-                event.status == Event.STATUS_OFFLINE_CLOSED
-                and ApplicationViewSet.check_hybrid_event_online_limit(event)
-            )
-            or (
-                event.status == Event.STATUS_ONLINE_CLOSED
-                and ApplicationViewSet.check_hybrid_event_offline_limit(event)
-            )
-        ):
-            event.status = Event.STATUS_CLOSED
-            logger.debug(
-                f"The status of event {event} was changed to {Event.STATUS_CLOSED}"
-            )
-        event.save()
-
     def perform_create(self, serializer):
         """
         Adds the user to the application if the request user is authenticated.
-        Triggers the authenticated user personal data update if the authenticated user
-        change personal data in the application.
+        Triggers the authenticated user's data updating if the user changes this data
+        in the application.
         Authomatically fills in the fields of an application of the authenticated user.
         Triggers creation of notification settings object if the user is anonymous.
         Triggers event participant limits checking and closure of registration
@@ -229,52 +114,10 @@ class ApplicationViewSet(CreateModelMixin, DestroyModelMixin, GenericViewSet):
             serializer.save()
             created_application = serializer.instance
             create_notification_settings(application=created_application)
-        ApplicationViewSet.check_event_limits_and_close_registration(
+        EventClosureController.check_event_limits_and_close_registration(
             serializer.validated_data["event"]
         )
 
-    # TODO: отменить заявку может только авторизованный (на входе в эндпойнт удаления
-    # стоит проверка в permission, наверно здесь можно дополнительно не проверять, что
-    # юзер авторизован);
-
-    # NO CHANGE ситуация 1: статус ивента "регистрация открыта", тогда не меняем статус
-
-    # OPEN ситуация 2: ивент имеет строгий формат и статус был "регистрация закрыта",
-    # тогда после отмены заявки меняем статус ивента на "регистрация открыта"
-
-    # ситуация 3: ивент имеет гибридный формат и статус "регистрация закрыта", тогда
-    # нужно посмотреть, какие лимиты имеет этот ивент (у него может не быть какого-то
-    # типа лимита, ведь эти поля необязательные) и какой формат был у отмененной заявки:
-
-    # 3.1 если у ивента были оба лимита и заявка была офлайн, тогда меняем статус ивента
-    # на "регистрация онлайн закрыта" (то есть возобновляем офлайн-регистрацию)
-    # 3.2 если у ивента были оба лимита и заявка была онлайн, тогда меняем статус ивента
-    # на "регистрация офлайн закрыта" (то есть возобновляем онлайн-регистрацию)
-    # 3.3 OPEN если у ивента был только офлайн-лимит и заявка была офлайн, то меняем
-    # статус на "регистрация открыта"
-    # 3.4 OPEN если у ивента был только онлайн-лимит и заявка была онлайн, то меняем
-    # статус на "регистрация открыта"
-    # 3.5 NO CHANGE если у ивента был только офлайн-лимит, а заявка была онлайн, то
-    # не меняем статус
-    # 3.6 NO CHANGE если у ивента был только онлайн-лимит, а заявка была офлайн, то
-    # не меняем статус
-    # 3.7 NO CHANGE если у ивента не было лимитов, то у него и статус должен был
-    # оставаться "регистрация открыта", но если админ в Админке вручную поменял статус
-    # на "регистрация закрыта", а потом кто-то отменил заявку, то не меняем статус
-    # ивента автоматически, пусть админ и дальше осуществляет ручное управление статусом
-    # ивента, раз он уже начал вмешиваться в автоматическую смену статусов
-
-    # ситуация 4: ивент имеет гибридный формат, офлайн-лимит и статус "регистрация
-    # офлайн закрыта":
-    # 4.1 OPEN заявка была офлайн, тогда меняем статус ивента на "регистрация открыта"
-    # 4.2 NO CHANGE заявка была онлайн, тогда не меняем статус ивента
-
-    # ситуация 5: ивент имеет гибридный формат, онлайн-лимит и статус "регистрация
-    # онлайн закрыта":
-    # 5.1 OPEN заявка была онлайн, тогда меняем статус ивента на "регистрация открыта"
-    # 5.2 NO CHANGE заявка была офлайн, тогда не меняем статус ивента
-
-    # TODO: add logging (event status changes)
     def perform_destroy(self, instance):
         """
         Deletes the application for participation in the event and re-opens
@@ -287,58 +130,9 @@ class ApplicationViewSet(CreateModelMixin, DestroyModelMixin, GenericViewSet):
             f"The application of user {self.request.user} to participate "
             f"in event {event} was deleted."
         )
-
-        # открываем регистрацию в целом
-        if event.format != Event.FORMAT_HYBRID and event.status == Event.STATUS_CLOSED:
-            event.status = Event.STATUS_OPEN  # ситуация 2
-        if (
-            event.format == Event.FORMAT_HYBRID
-            and event.participant_offline_limit
-            and event.status == Event.STATUS_OFFLINE_CLOSED
-            and application_format == Event.FORMAT_OFFLINE
-        ):
-            event.status = Event.STATUS_OPEN  # ситуация 4.1
-        if (
-            event.format == Event.FORMAT_HYBRID
-            and event.participant_online_limit
-            and event.status == Event.STATUS_ONLINE_CLOSED
-            and application_format == Event.FORMAT_ONLINE
-        ):
-            event.status = Event.STATUS_OPEN  # ситуация 5.1
-        if (
-            event.format == Event.FORMAT_HYBRID
-            and event.participant_offline_limit
-            and not event.participant_online_limit
-            and event.status == Event.STATUS_CLOSED
-            and application_format == Event.FORMAT_OFFLINE
-        ):
-            event.status = Event.STATUS_OPEN  # ситуация 3.3
-        if (
-            event.format == Event.FORMAT_HYBRID
-            and not event.participant_offline_limit
-            and event.participant_online_limit
-            and event.status == Event.STATUS_CLOSED
-            and application_format == Event.FORMAT_ONLINE
-        ):
-            event.status = Event.STATUS_OPEN  # ситуация 3.4
-        if (
-            event.format == Event.FORMAT_HYBRID
-            and event.participant_offline_limit
-            and event.participant_online_limit
-            and event.status == Event.STATUS_CLOSED
-            and application_format == Event.FORMAT_ONLINE
-        ):
-            event.status = Event.STATUS_OFFLINE_CLOSED  # ситуация 3.1
-        if (
-            event.format == Event.FORMAT_HYBRID
-            and event.participant_offline_limit
-            and event.participant_online_limit
-            and event.status == Event.STATUS_CLOSED
-            and application_format == Event.FORMAT_OFFLINE
-        ):
-            event.status = Event.STATUS_ONLINE_CLOSED  # ситуация 3.2
-
-        event.save()
+        EventReopeningController.check_event_limits_and_reopen_registration(
+            event, application_format
+        )
 
 
 class NotificationSettingsViewSet(RetrieveModelMixin, UpdateModelMixin, GenericViewSet):
